requirements.txt
django = 2.2
requests
django-pure-pagination
pillow

